refactor CliSyntax into static inner classes
