+++ src/main/java/seedu/jarvis/logic/parser/CliSyntax.java
@@ -4,18 +4,26 @@ package seedu.jarvis.logic.parser;
  * Contains Command Line Interface (CLI) syntax definitions common to multiple commands
  */
 public class CliSyntax {
+    /**
+     * CLI syntax for AddressBook commands.
+     */
+    public static class AddressSyntax {
+        public static final Prefix PREFIX_TASK_TYPE = new Prefix("t/");
+        public static final Prefix PREFIX_TASK_DES = new Prefix("des/");
+        public static final Prefix PREFIX_DATE = new Prefix("d/");
+        public static final Prefix PREFIX_PRIORITY = new Prefix("p/");
+        public static final Prefix PREFIX_FREQ = new Prefix("f/");
+        public static final Prefix PREFIX_NAME = new Prefix("n/");
+        public static final Prefix PREFIX_PHONE = new Prefix("p/");
+        public static final Prefix PREFIX_EMAIL = new Prefix("e/");
+        public static final Prefix PREFIX_ADDRESS = new Prefix("a/");
+        public static final Prefix PREFIX_TAG = new Prefix("t/");
+    }
 
-    /* Prefix definitions */
-    public static final Prefix PREFIX_TASK_TYPE = new Prefix("t/");
-    public static final Prefix PREFIX_TASK_DES = new Prefix("des/");
-    public static final Prefix PREFIX_DATE = new Prefix("d/");
-    public static final Prefix PREFIX_PRIORITY = new Prefix("p/");
-    public static final Prefix PREFIX_FREQ = new Prefix("f/");
-    public static final Prefix PREFIX_NAME = new Prefix("n/");
-    public static final Prefix PREFIX_PHONE = new Prefix("p/");
-    public static final Prefix PREFIX_EMAIL = new Prefix("e/");
-    public static final Prefix PREFIX_ADDRESS = new Prefix("a/");
-    public static final Prefix PREFIX_TAG = new Prefix("t/");
-    public static final Prefix PREFIX_UNDO_REDO = new Prefix("r/");
-
+    /**
+     * CLI syntax for Undo/Redo commands.
+     */
+    public static class UndoRedoSyntax {
+        public static final Prefix PREFIX_UNDO_REDO = new Prefix("r/");
+    }
 }
