+++ src/test/java/seedu/jarvis/logic/commands/CommandTestUtil.java
@@ -2,11 +2,11 @@ package seedu.jarvis.logic.commands;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.jarvis.logic.parser.CliSyntax.PREFIX_ADDRESS;
-import static seedu.jarvis.logic.parser.CliSyntax.PREFIX_EMAIL;
-import static seedu.jarvis.logic.parser.CliSyntax.PREFIX_NAME;
-import static seedu.jarvis.logic.parser.CliSyntax.PREFIX_PHONE;
-import static seedu.jarvis.logic.parser.CliSyntax.PREFIX_TAG;
+import static seedu.jarvis.logic.parser.CliSyntax.AddressSyntax.PREFIX_ADDRESS;
+import static seedu.jarvis.logic.parser.CliSyntax.AddressSyntax.PREFIX_EMAIL;
+import static seedu.jarvis.logic.parser.CliSyntax.AddressSyntax.PREFIX_NAME;
+import static seedu.jarvis.logic.parser.CliSyntax.AddressSyntax.PREFIX_PHONE;
+import static seedu.jarvis.logic.parser.CliSyntax.AddressSyntax.PREFIX_TAG;
 import static seedu.jarvis.testutil.Assert.assertThrows;
 
 import java.util.ArrayList;
